Created a query and integrated the response in Student Page //mergeconflict

+++ Review.Frontend/src/api/ReviewApi.js
@@ -9,32 +9,37 @@ const apiClient = axios.create({
     },
 });
 
-export const CreateQuestionnaires = async () => {
-    const response = await apiClient.post(`/questionnaires`);
-    return response.data;
-}
-
-export const GetQuestionnaireSummary = async (questionnaireId) => {
-    const response = await apiClient.get(`/questionnaires/${questionnaireId}`)
-    return response.data;
-}
-
-export const GetEvaluation = async (evaluationId) => {
-    const response = await apiClient.get(`/evaluations/${evaluationId}`)
-    return response.data;
-}
-
-export const UpdateEvaluation = async (evaluationId) => {
-    const response = await apiClient.patch(`/evaluations/${evaluationId}`)
-    return response.data;
-}
-
-export const DeleteQuestionnaire = async (questionnaireId) => {
-    const response = await apiClient.delete(`/questionnaires/${questionnaireId}`)
-    return response.data;
-}
-
 export const LoginWithGoogle = async (idToken) => {
     const { data } = await apiClient.post('/auth/google', { IdToken: idToken });
     return data;
 };
+
+export const GetStudentByEmail = async (email) => {
+    const {data} =await apiClient.post('/students/context' ,{email});
+    return data;
+}
+
+export const CreateQuestionnaires = async () => {
+    const response = await apiClient.post(`/questionnaires`); //
+    return response.data;
+}
+
+export const GetQuestionnaireSummary = async (questionnaireId) => {
+    const response = await apiClient.get(`/questionnaires/${questionnaireId}`) //globalisan az osszes statisztika admin keri le
+    return response.data;
+}
+
+export const GetEvaluation = async (evaluationId) => {
+    const response = await apiClient.get(`/evaluations/${evaluationId}`)  //minden tanarnak a sajat tantargya ertekelese
+    return response.data;
+}
+
+export const UpdateEvaluation = async (evaluationId) => {
+    const response = await apiClient.patch(`/evaluations/${evaluationId}`) //minden questionnal
+    return response.data;
+}
+
+export const DeleteQuestionnaire = async (questionnaireId) => {
+    const response = await apiClient.delete(`/questionnaires/${questionnaireId}`) //admin deletel mindent a db bol
+    return response.data;
+}
